refactor: separate integration test files
